% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings.R
\name{elo}
\alias{elo}
\title{Get Elo Rating}
\usage{
elo(id)
}
\arguments{
\item{id}{A USCF ID number}
}
\value{
a classical Elo rating, as an integer
}
\description{
Given a USCF ID number, the elo() function will return the classical Elo rating of that player.
}
\examples{
elo(12641216)
elo(13145890)
}
